requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
duckdb==0.10.2
exceptiongroup==1.2.0
fastapi==0.110.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
palmerpenguins==0.1.4
pandas==2.2.2
pins==0.8.5
pip-tools==7.4.1
plotly==5.21.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsconnect_python==1.23.0
scikit-learn==1.4.2
scipy==1.13.0
semver==2.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==8.2.3
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vetiver==0.2.4
xxhash==3.4.1
zipp==3.18.1
